serverchatbot/requirements.txt
flask==3.1.2
pyngrok==7.4.1
transformers==4.46.2
torch==2.5.1
torchvision==0.20.1
pillow==11.0.0
accelerate==1.1.1
blinker==1.9.0
psutil==6.0.0
requests==2.32.3
bitsandbytes==0.43.1
git+https://github.com/huggingface/transformers
